Node/acct: switch watcher to use tm library
